+++ CHANGELOG.md
@@ -8,3 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - initial forked code
+
+### Fixed
+- calculation of nesting increments